Add support for navigating a looping path
This support uses the new nav2_client_loop action that supports
specifying a goal point and an intermediate waypoint.  This
action expects the navigation path plannter to be the custom
nav2_fixed_path_planner which has been configured with the
desired waypoints in the navigation_fixed_path_pure_pursuit.yaml
file.  (Future support will allow setting the waypoints via the
behavior tree.)

The intended initial use case is for driving the Elsabot Jeep
powerwheels in a loop in the backyard.

The full behavior tree will support voice command start/stop
looping.

This commit disables support for the XArm robot arm for now since
it isn't used.
FIX - revise the makefile to make enabling this support cleaner.