Publish goal_update from BT when following
The /goal_update topic should only be published when in following mode.  The tracker node will no longer publish goal updates.  The BT will now publish goal updates via the new PublishPositionAsGoalAction when it is actively in the following mode.  The ObjectTrackerStatusAction action was revised to provide the tracked position as an output.